moved tests to separate file and dir
